kernel: import RMX1901 4.14 r011 source snapshot
Publish the exact source tree used by B14-M03-r011 without importing the unrelated 810594-commit Qualcomm/OpenELA/Lineage history into the RMX1901 repository.

Original-Commit: 2ffb7ab18194fbb81220600e0679dd116cdffd91

Original-Tree: f354992746e972d87af86adac6db4a0283a673dc

Base-Remote-Commit: e7a1612ee260c815354c80c13440eaf49e8b30f4

Vmlinux-Build-ID: 59821879144920cb

+++ include/linux/slimbus/slimbus.h
@@ -514,7 +514,7 @@ enum slim_clk_state {
  * @devs: List of devices on this controller
  * @wq: Workqueue per controller used to notify devices when they report present
  * @txnt: Table of transactions having transaction ID
- * @last_tid: size of the table txnt (can't grow beyond 256 since TID is 8-bits)
+ * @last_tid: size of the table txnt (can't grow beyond 256)
  * @ports: Ports associated with this controller
  * @nports: Number of ports supported by the controller
  * @chans: Channels associated with this controller
@@ -576,7 +576,7 @@ struct slim_controller {
 	struct list_head	devs;
 	struct workqueue_struct *wq;
 	struct slim_msg_txn	*txnt[SLIM_MAX_TXNS];
-	u8			last_tid;
+	unsigned int		last_tid;
 	spinlock_t		txn_lock;
 	struct slim_port	*ports;
 	int			nports;
@@ -726,7 +726,7 @@ struct slim_boardinfo {
 
 /*
  * slim_get_logical_addr: Return the logical address of a slimbus device.
- * @sb: client handle requesting the adddress.
+ * @sb: client handle requesting the address.
  * @e_addr: Elemental address of the device.
  * @e_len: Length of e_addr
  * @laddr: output buffer to store the address
